remove-action-from-application-contoroller
application-contorollerからメソッドを切り出し、controllers/consernに移動
